Maintenances of Forms

+++ app/Http/Controllers/AcademicDevelopment/ReferenceController.php
@@ -2,15 +2,16 @@
 
 namespace App\Http\Controllers\AcademicDevelopment;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use App\Models\Reference;
-use App\Models\ReferenceDocument;
+use Illuminate\Http\Request;
 use App\Models\TemporaryFile;
-use Illuminate\Support\Facades\Storage;
+use App\Models\ReferenceDocument;
 use Illuminate\Support\Facades\DB;
 use App\Models\Maintenance\College;
+use App\Http\Controllers\Controller;
 use App\Models\Maintenance\Department;
+use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\AcademicDevelopmentForm;
 
 class ReferenceController extends Controller
 {
@@ -36,6 +37,8 @@ class ReferenceController extends Controller
      */
     public function create()
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $colleges = College::all();
@@ -50,6 +53,9 @@ class ReferenceController extends Controller
      */
     public function store(Request $request)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
+      
         $request->validate([
             'category' => 'required',
             'level' => 'required',
@@ -112,6 +118,8 @@ class ReferenceController extends Controller
      */
     public function show(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
 
         $category = DB::select("CALL get_dropdown_name_by_id(".$rtmmi->category.")");
@@ -133,6 +141,8 @@ class ReferenceController extends Controller
      */
     public function edit(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
@@ -160,6 +170,9 @@ class ReferenceController extends Controller
      */
     public function update(Request $request, Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
+      
         $request->validate([
             'category' => 'required',
             'level' => 'required',
@@ -222,6 +235,8 @@ class ReferenceController extends Controller
      */
     public function destroy(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $rtmmi->delete();
         ReferenceDocument::where('reference_id', $rtmmi->id)->delete();
 
@@ -235,6 +250,8 @@ class ReferenceController extends Controller
     }
 
     public function removeDoc($filename){
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         ReferenceDocument::where('filename', $filename)->delete();
         // Storage::delete('documents/'.$filename);
         return true;
