+++ app/Http/Controllers/AcademicDevelopment/StudentTrainingController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use App\Models\StudentTrainingDocument;
 use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\AcademicDevelopmentForm;
 
 class StudentTrainingController extends Controller
 {
@@ -31,6 +32,8 @@ class StudentTrainingController extends Controller
      */
     public function create()
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(4)");
 
         return view('academic-development.student-training.create', compact('studentFields'));
@@ -44,6 +47,8 @@ class StudentTrainingController extends Controller
      */
     public function store(Request $request)
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $student_training = StudentTraining::create($input);
@@ -83,6 +88,8 @@ class StudentTrainingController extends Controller
      */
     public function show(StudentTraining $student_training)
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(4)");
 
         $documents = StudentTrainingDocument::where('student_training_id', $student_training->id)->get()->toArray();
@@ -101,6 +108,8 @@ class StudentTrainingController extends Controller
      */
     public function edit(StudentTraining $student_training)
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(4)");
 
         $documents = StudentTrainingDocument::where('student_training_id', $student_training->id)->get()->toArray();
@@ -119,6 +128,8 @@ class StudentTrainingController extends Controller
      */
     public function update(Request $request, StudentTraining $student_training)
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $student_training->update($input);
@@ -157,12 +168,16 @@ class StudentTrainingController extends Controller
      */
     public function destroy(StudentTraining $student_training)
     {
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         StudentTrainingDocument::where('student_training_id', $student_training->id)->delete();
         $student_training->delete();
         return redirect()->route('student-training.index')->with('student_success', 'Your accomplishment in Student Attended Seminars and Trainings has been deleted.');
     }
 
     public function removeDoc($filename){
+        if(AcademicDevelopmentForm::where('id', 4)->pluck('is_active')->first() == 0)
+            return view('inactive');
         StudentTrainingDocument::where('filename', $filename)->delete();
         return true;
     }
