+++ app/Http/Controllers/ExtensionPrograms/OutreachProgramController.php
@@ -10,6 +10,7 @@ use App\Models\PartnershipDocument;
 use App\Http\Controllers\Controller;
 use App\Models\OutreachProgramDocument;
 use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\ExtensionProgramForm;
 
 class OutreachProgramController extends Controller
 {
@@ -31,6 +32,8 @@ class OutreachProgramController extends Controller
      */
     public function create()
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $outreachFields = DB::select("CALL get_extension_program_fields_by_form_id('7')");
 
         return view('extension-programs.outreach-program.create', compact('outreachFields'));
@@ -44,6 +47,8 @@ class OutreachProgramController extends Controller
      */
     public function store(Request $request)
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $outreach = OutreachProgram::create($input);
@@ -84,6 +89,8 @@ class OutreachProgramController extends Controller
      */
     public function show(OutreachProgram $outreach_program)
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $outreachFields = DB::select("CALL get_extension_program_fields_by_form_id('7')");
 
         $values = $outreach_program->toArray();
@@ -101,6 +108,8 @@ class OutreachProgramController extends Controller
      */
     public function edit(OutreachProgram $outreach_program)
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $outreachFields = DB::select("CALL get_extension_program_fields_by_form_id('7')");
 
         $values = $outreach_program->toArray();
@@ -119,6 +128,8 @@ class OutreachProgramController extends Controller
      */
     public function update(Request $request, OutreachProgram $outreach_program)
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $outreach_program->update($input);
@@ -158,12 +169,16 @@ class OutreachProgramController extends Controller
      */
     public function destroy(OutreachProgram $outreach_program)
     {
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         OutreachProgramDocument::where('outreach_program_id', $outreach_program->id)->delete();
         $outreach_program->delete();
         return redirect()->route('outreach-program.index')->with('outreach_success', 'Your accomplishment in Community Relations and Outreach Program has been deleted.');
     }
 
     public function removeDoc($filename){
+        if(ExtensionProgramForm::where('id', 7)->pluck('is_active')->first() == 0)
+            return view('inactive');
         OutreachProgramDocument::where('filename', $filename)->delete();
         return true;
     }
